Refactored classes in own file.
